HBASE-7692 Orderly: 32-bit Integer types
This commit add support for variable and fixed-width serialization of
both signed and unsigned 32-bit Integer types. The class hierarchy
follows two ancestries.

Fixed-with serialization formats are implemented by or extend
FixedIntWritableRowKey. This produces a 4-byte, fixed-width sortable
format based on IntWritable. Support for unsigned IntWritables and
native Integer types are derived from this class.

Variable-width serialization is based on AbstractVarIntRowKey. This
class handles the encoding wrapper, primarily management of a
header-byte. Subclasses implement the details specific to signed and
unsigned IntWritables and native Integers.

Also includes and example.